+++ application-infrastructure/src/main/kotlin/hs/kr/equus/application/global/document/pdf/config/Font.kt
@@ -1,4 +1,4 @@
-package hs.kr.equus.application.global.util.pdf
+package hs.kr.equus.application.global.document.pdf.config
 
 object Font {
     val fonts = listOf(
