+++ application-infrastructure/src/main/kotlin/hs/kr/equus/application/global/document/pdf/generator/PdfProcessor.kt
@@ -1,10 +1,11 @@
-package hs.kr.equus.application.global.util.pdf
+package hs.kr.equus.application.global.document.pdf.generator
 
 import com.itextpdf.html2pdf.HtmlConverter
 import com.itextpdf.kernel.pdf.PdfDocument
 import com.itextpdf.kernel.pdf.PdfReader
 import com.itextpdf.kernel.pdf.PdfWriter
 import com.itextpdf.kernel.utils.PdfMerger
+import hs.kr.equus.application.global.document.pdf.config.ConverterPropertiesCreator
 import org.springframework.stereotype.Component
 import java.io.ByteArrayInputStream
 import java.io.ByteArrayOutputStream
